web/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.5
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
idna==3.6
mysqlclient==2.1.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.9.0
ua-parser==0.18.0
urllib3==2.1.0
uWSGI==2.0.22